Split concrete classes and interfaces
